Renamed file GOPath
