+++ src/core/go_path.h
